+++ CHANGELOG.md
@@ -3,4 +3,5 @@
 ## v0.3.0
 
 * (C) Extracted from Tideland Go Library as part of split
+* (C) Reimplemented stopwatch